src/api-service-shivas/requirements.txt
google-cloud-storage
fastapi
starlette
bcrypt
pydantic
pyjwt
httpx
pytest
pytest-cov